Get very basic variable declaration working
This required implementing an amount of infrastructure for handling of
lexical scopes and symbol resolution using them. Also some refinement of
existing APIs given further learning, and an `is raw` trait support in
the AST compiler so we can properly handle Scalar containers without
them evaporating. This is still highly incomplete; it's just enough to
get an initial test to pass for variable declaration, and now much more
needs to be filled out.